Remove dashboard2 app
