--- meinberlin/apps/contrib/templates/meinberlin_contrib/includes/inline_form.html
@@ -1,23 +1,0 @@
-{% load i18n %}
-<div class="js-form formset__form form-group">
-    {% if can_delete %}
-        <input type="checkbox" name="{{ form.DELETE.html_name }}" id="{{ form.DELETE.id_for_label }}" class="formset__delete">
-    {% endif %}
-    <label for="{{ form.name.id_for_label }}" class="sr-only">
-        {{ form.name.label }}
-    </label>
-    {{ form.name }}
-    {% if not form.name.value %}
-        <button type="button" class="formset__delete-label js-remove-form" title="{% trans 'Remove category' %}">
-            <i class="fa fa-times" aria-label="{% trans 'Remove category' %}"></i>
-        </button>
-    {% elif can_delete %}
-        <label for="{{ form.DELETE.id_for_label }}" class="formset__delete-label" title="{% trans 'Remove category' %}">
-            <i class="fa fa-trash" aria-label="{% trans 'Remove category' %}"></i>
-        </label>
-    {% endif %}
-    {% for hidden in form.hidden_fields %}
-        {{ hidden }}
-    {% endfor %}
-    {{ form.name.errors }}
-</div>
