+++ meinberlin/apps/dashboard/urls.py
@@ -36,6 +36,13 @@ urlpatterns = [
         r'/plans/(?P<pk>\d+)/delete/$',
         plan_views.DashboardPlanDeleteView.as_view(),
         name='plan-delete'),
+    url(r'^projects/(?P<project_slug>[-\w_]+)/blueprints/$',
+        views.ModuleBlueprintListView.as_view(),
+        name='module-blueprint-list'),
+    url(r'^projects/(?P<project_slug>[-\w_]+)/blueprints/'
+        '(?P<blueprint_slug>[-\w_]+)/$',
+        views.ModuleCreateView.as_view(),
+        name='module-create'),
 
     # Overwrite adhocracy4 core urls with meinBerlin urls
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/blueprints/'
