--- meinberlin/apps/dashboard2/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class Config(AppConfig):
-    name = 'meinberlin.apps.dashboard2'
-    label = 'meinberlin_dashboard2'
