--- meinberlin/apps/dashboard2/templates/meinberlin_dashboard2/includes/module_categories_form.html
@@ -1,22 +1,0 @@
-{% load i18n %}
-
-<p class="form-hint">{% blocktrans %}In this section, you can create
-categories. If you create any, users must assign their contributions
-to one of them. This way, content can be classified.{% endblocktrans %}</p>
-
-<div class="js-formset formset"
-     data-prefix="{{ form.prefix }}">
-    {{ form.management_form }}
-
-    {% for category_form in form %}
-        {% include 'meinberlin_contrib/includes/inline_form.html' with form=category_form can_delete=form.can_delete %}
-    {% endfor %}
-    <div class="js-form-template formset__template">
-        {% include 'meinberlin_contrib/includes/inline_form.html' with form=form.empty_form can_delete=form.can_delete %}
-    </div>
-
-    <button class="btn btn--secondary btn--small btn--full js-add-form" type="button">
-        <i class="fa fa-plus" aria-hidden="true"></i>
-        {% trans 'Add category' %}
-    </button>
-</div>
