--- meinberlin/apps/dashboard2/templates/meinberlin_dashboard2/includes/removeable_user_list.html
@@ -1,30 +1,0 @@
-{% load i18n %}
-
-{% if users|length %}
-    <table>
-        <thead>
-            <tr>
-                <th>{% trans 'Name' %}</th>
-                <th>{% trans 'Email' %}</th>
-                <th><span class="sr-only">{% trans 'Remove' %}</span></th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for user in users %}
-                <tr>
-                    <td>{{ user.username }}</td>
-                    <td>{{ user.email }}</td>
-                    <td class="td--action">
-                        <form method="post">
-                            {% csrf_token %}
-                            <input type="hidden" name="user_pk" value="{{ user.pk }}">
-                            <button type="submit" name="submit_action" value="remove_user" title="{% trans 'Remove' %}">
-                                <i class="fa fa-times" aria-label="{% trans 'Remove' %}"></i>
-                            </button>
-                        </form>
-                    </td>
-                </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-{% endif %}
