--- meinberlin/apps/dashboard2/templates/meinberlin_dashboard2/includes/users_from_email_form.html
@@ -1,22 +1,0 @@
-{% load i18n widget_tweaks %}
-
-<form method="POST">
-    {% csrf_token %}
-    <div class="form-group">
-        <label for="{{ form.add_users.id_for_label }}">
-            {{ form.add_users.label }}
-        </label>
-        {% if form.add_users.help_text %}
-        <div class="form-hint">
-            {{ form.add_users.help_text }}
-        </div>
-        {% endif %}
-        <div class="input-group">
-            <div class="input-group__input widget widget--{{ form.add_users|widget_type }}">
-                {{ form.add_users }}
-            </div>
-            <input type="submit" class="btn btn--primary input-group__after" value="{% trans 'Add' %}"/>
-        </div>
-        {{ form.add_users.errors }}
-    </div>
-</form>
