--- meinberlin/apps/dashboard2/urls.py
@@ -1,14 +1,0 @@
-from django.conf.urls import url
-
-from . import views
-
-app_name = 'a4dashboard'
-urlpatterns = [
-    url(r'^projects/(?P<project_slug>[-\w_]+)/blueprints/$',
-        views.ModuleBlueprintListView.as_view(),
-        name='module-blueprint-list'),
-    url(r'^projects/(?P<project_slug>[-\w_]+)/blueprints/'
-        '(?P<blueprint_slug>[-\w_]+)/$',
-        views.ModuleCreateView.as_view(),
-        name='module-create'),
-]
